# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriControl - Nutritional tracking app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileSex = (typeof ProfileSex)[keyof typeof ProfileSex];

export const ProfileSex = {
  M: "M",
  F: "F",
} as const;

export type ProfileActivityLevel =
  (typeof ProfileActivityLevel)[keyof typeof ProfileActivityLevel];

export const ProfileActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
  extra_active: "extra_active",
} as const;

export type ProfileCaloricGoal =
  (typeof ProfileCaloricGoal)[keyof typeof ProfileCaloricGoal];

export const ProfileCaloricGoal = {
  maintain: "maintain",
  lose_mild: "lose_mild",
  lose_moderate: "lose_moderate",
  lose_aggressive: "lose_aggressive",
  gain: "gain",
} as const;

export interface Profile {
  id: number;
  name: string;
  age: number;
  sex: ProfileSex;
  weightKg: number;
  heightCm: number;
  activityLevel: ProfileActivityLevel;
  caloricGoal: ProfileCaloricGoal;
  bmi: number;
  bmr: number;
  tdee: number;
  targetCalories: number;
  targetProteinG: number;
  targetCarbsG: number;
  targetFatG: number;
  targetWaterMl: number;
  intermittentFasting: boolean;
  fastingStartHour?: number;
  fastingEndHour?: number;
}

export type UpdateProfileBodySex =
  (typeof UpdateProfileBodySex)[keyof typeof UpdateProfileBodySex];

export const UpdateProfileBodySex = {
  M: "M",
  F: "F",
} as const;

export type UpdateProfileBodyActivityLevel =
  (typeof UpdateProfileBodyActivityLevel)[keyof typeof UpdateProfileBodyActivityLevel];

export const UpdateProfileBodyActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
  extra_active: "extra_active",
} as const;

export type UpdateProfileBodyCaloricGoal =
  (typeof UpdateProfileBodyCaloricGoal)[keyof typeof UpdateProfileBodyCaloricGoal];

export const UpdateProfileBodyCaloricGoal = {
  maintain: "maintain",
  lose_mild: "lose_mild",
  lose_moderate: "lose_moderate",
  lose_aggressive: "lose_aggressive",
  gain: "gain",
} as const;

export interface UpdateProfileBody {
  name?: string;
  age?: number;
  sex?: UpdateProfileBodySex;
  weightKg?: number;
  heightCm?: number;
  activityLevel?: UpdateProfileBodyActivityLevel;
  caloricGoal?: UpdateProfileBodyCaloricGoal;
  intermittentFasting?: boolean;
  fastingStartHour?: number;
  fastingEndHour?: number;
}

export interface Food {
  id: number;
  name: string;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
  fiberG?: number;
  servingG: number;
}

export type DiaryEntryMealType =
  (typeof DiaryEntryMealType)[keyof typeof DiaryEntryMealType];

export const DiaryEntryMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface DiaryEntry {
  id: number;
  date: string;
  mealType: DiaryEntryMealType;
  foodId?: number;
  foodName: string;
  quantityG: number;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}

export type CreateDiaryEntryBodyMealType =
  (typeof CreateDiaryEntryBodyMealType)[keyof typeof CreateDiaryEntryBodyMealType];

export const CreateDiaryEntryBodyMealType = {
  breakfast: "breakfast",
  lunch: "lunch",
  dinner: "dinner",
  snack: "snack",
} as const;

export interface CreateDiaryEntryBody {
  date: string;
  mealType: CreateDiaryEntryBodyMealType;
  foodId?: number;
  foodName: string;
  quantityG: number;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}

export interface DiaryDay {
  date: string;
  entries: DiaryEntry[];
  totalCalories: number;
  totalProteinG: number;
  totalCarbsG: number;
  totalFatG: number;
}

export interface Hydration {
  date: string;
  totalMl: number;
  cups: number;
}

export interface AddHydrationBody {
  date: string;
  amountMl: number;
}

export interface Recipe {
  id: number;
  name: string;
  description?: string;
  servings: number;
  caloriesPerServing: number;
  proteinGPerServing: number;
  carbsGPerServing: number;
  fatGPerServing: number;
  ingredients: string[];
  instructions?: string;
}

export interface CreateRecipeBody {
  name: string;
  description?: string;
  servings: number;
  caloriesPerServing: number;
  proteinGPerServing: number;
  carbsGPerServing: number;
  fatGPerServing: number;
  ingredients: string[];
  instructions?: string;
}

export interface WeightEntry {
  id: number;
  date: string;
  weightKg: number;
}

export interface LogWeightBody {
  date: string;
  weightKg: number;
}

export interface DailySummary {
  date: string;
  totalCalories: number;
  targetCalories: number;
}

export interface ProgressHistory {
  weightEntries: WeightEntry[];
  dailySummaries: DailySummary[];
}

export type GetDiaryEntriesParams = {
  date: string;
};

export type SearchFoodsParams = {
  q: string;
  limit?: number;
};

export type GetHydrationParams = {
  date: string;
};

export type ResetHydrationBody = {
  date: string;
};