# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriControl - Nutritional tracking app API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get user profile
 */
export const GetProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  age: zod.number(),
  sex: zod.enum(["M", "F"]),
  weightKg: zod.number(),
  heightCm: zod.number(),
  activityLevel: zod.enum([
    "sedentary",
    "lightly_active",
    "moderately_active",
    "very_active",
    "extra_active",
  ]),
  caloricGoal: zod.enum([
    "maintain",
    "lose_mild",
    "lose_moderate",
    "lose_aggressive",
    "gain",
  ]),
  bmi: zod.number(),
  bmr: zod.number(),
  tdee: zod.number(),
  targetCalories: zod.number(),
  targetProteinG: zod.number(),
  targetCarbsG: zod.number(),
  targetFatG: zod.number(),
  targetWaterMl: zod.number(),
  intermittentFasting: zod.boolean(),
  fastingStartHour: zod.number().optional(),
  fastingEndHour: zod.number().optional(),
});

/**
 * @summary Update user profile
 */
export const UpdateProfileBody = zod.object({
  name: zod.string().optional(),
  age: zod.number().optional(),
  sex: zod.enum(["M", "F"]).optional(),
  weightKg: zod.number().optional(),
  heightCm: zod.number().optional(),
  activityLevel: zod
    .enum([
      "sedentary",
      "lightly_active",
      "moderately_active",
      "very_active",
      "extra_active",
    ])
    .optional(),
  caloricGoal: zod
    .enum(["maintain", "lose_mild", "lose_moderate", "lose_aggressive", "gain"])
    .optional(),
  intermittentFasting: zod.boolean().optional(),
  fastingStartHour: zod.number().optional(),
  fastingEndHour: zod.number().optional(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  age: zod.number(),
  sex: zod.enum(["M", "F"]),
  weightKg: zod.number(),
  heightCm: zod.number(),
  activityLevel: zod.enum([
    "sedentary",
    "lightly_active",
    "moderately_active",
    "very_active",
    "extra_active",
  ]),
  caloricGoal: zod.enum([
    "maintain",
    "lose_mild",
    "lose_moderate",
    "lose_aggressive",
    "gain",
  ]),
  bmi: zod.number(),
  bmr: zod.number(),
  tdee: zod.number(),
  targetCalories: zod.number(),
  targetProteinG: zod.number(),
  targetCarbsG: zod.number(),
  targetFatG: zod.number(),
  targetWaterMl: zod.number(),
  intermittentFasting: zod.boolean(),
  fastingStartHour: zod.number().optional(),
  fastingEndHour: zod.number().optional(),
});

/**
 * @summary Get diary entries for a date
 */
export const GetDiaryEntriesQueryParams = zod.object({
  date: zod.date(),
});

export const GetDiaryEntriesResponse = zod.object({
  date: zod.coerce.date(),
  entries: zod.array(
    zod.object({
      id: zod.number(),
      date: zod.coerce.date(),
      mealType: zod.enum(["breakfast", "lunch", "dinner", "snack"]),
      foodId: zod.number().optional(),
      foodName: zod.string(),
      quantityG: zod.number(),
      calories: zod.number(),
      proteinG: zod.number(),
      carbsG: zod.number(),
      fatG: zod.number(),
    }),
  ),
  totalCalories: zod.number(),
  totalProteinG: zod.number(),
  totalCarbsG: zod.number(),
  totalFatG: zod.number(),
});

/**
 * @summary Add a food entry to diary
 */
export const AddDiaryEntryBody = zod.object({
  date: zod.coerce.date(),
  mealType: zod.enum(["breakfast", "lunch", "dinner", "snack"]),
  foodId: zod.number().optional(),
  foodName: zod.string(),
  quantityG: zod.number(),
  calories: zod.number(),
  proteinG: zod.number(),
  carbsG: zod.number(),
  fatG: zod.number(),
});

/**
 * @summary Delete a diary entry
 */
export const DeleteDiaryEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Search TACO food table
 */
export const searchFoodsQueryLimitDefault = 20;

export const SearchFoodsQueryParams = zod.object({
  q: zod.coerce.string(),
  limit: zod.coerce.number().default(searchFoodsQueryLimitDefault),
});

export const SearchFoodsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  calories: zod.number(),
  proteinG: zod.number(),
  carbsG: zod.number(),
  fatG: zod.number(),
  fiberG: zod.number().optional(),
  servingG: zod.number(),
});
export const SearchFoodsResponse = zod.array(SearchFoodsResponseItem);

/**
 * @summary Get hydration for a date
 */
export const GetHydrationQueryParams = zod.object({
  date: zod.date(),
});

export const GetHydrationResponse = zod.object({
  date: zod.coerce.date(),
  totalMl: zod.number(),
  cups: zod.number(),
});

/**
 * @summary Add water intake
 */
export const AddHydrationBody = zod.object({
  date: zod.coerce.date(),
  amountMl: zod.number(),
});

export const AddHydrationResponse = zod.object({
  date: zod.coerce.date(),
  totalMl: zod.number(),
  cups: zod.number(),
});

/**
 * @summary Reset hydration for a date
 */
export const ResetHydrationBody = zod.object({
  date: zod.coerce.date(),
});

export const ResetHydrationResponse = zod.object({
  date: zod.coerce.date(),
  totalMl: zod.number(),
  cups: zod.number(),
});

/**
 * @summary Get all recipes
 */
export const GetRecipesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  servings: zod.number(),
  caloriesPerServing: zod.number(),
  proteinGPerServing: zod.number(),
  carbsGPerServing: zod.number(),
  fatGPerServing: zod.number(),
  ingredients: zod.array(zod.string()),
  instructions: zod.string().optional(),
});
export const GetRecipesResponse = zod.array(GetRecipesResponseItem);

/**
 * @summary Create a recipe
 */
export const CreateRecipeBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  servings: zod.number(),
  caloriesPerServing: zod.number(),
  proteinGPerServing: zod.number(),
  carbsGPerServing: zod.number(),
  fatGPerServing: zod.number(),
  ingredients: zod.array(zod.string()),
  instructions: zod.string().optional(),
});

/**
 * @summary Get recipe details
 */
export const GetRecipeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRecipeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().optional(),
  servings: zod.number(),
  caloriesPerServing: zod.number(),
  proteinGPerServing: zod.number(),
  carbsGPerServing: zod.number(),
  fatGPerServing: zod.number(),
  ingredients: zod.array(zod.string()),
  instructions: zod.string().optional(),
});

/**
 * @summary Delete a recipe
 */
export const DeleteRecipeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get progress history (last 30 days)
 */
export const GetProgressResponse = zod.object({
  weightEntries: zod.array(
    zod.object({
      id: zod.number(),
      date: zod.coerce.date(),
      weightKg: zod.number(),
    }),
  ),
  dailySummaries: zod.array(
    zod.object({
      date: zod.coerce.date(),
      totalCalories: zod.number(),
      targetCalories: zod.number(),
    }),
  ),
});

/**
 * @summary Log a weight entry
 */
export const LogWeightBody = zod.object({
  date: zod.coerce.date(),
  weightKg: zod.number(),
});

/**
 * @summary Get weekly diary summary (calories per day for the last 7 days)
 */
export const GetDiarySummaryResponseItem = zod.object({
  date: zod.coerce.date(),
  totalCalories: zod.number(),
  targetCalories: zod.number(),
});
export const GetDiarySummaryResponse = zod.array(GetDiarySummaryResponseItem);