# lib/api-zod/src/generated/types/diaryEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriControl - Nutritional tracking app API
 * OpenAPI spec version: 0.1.0
 */
import type { DiaryEntryMealType } from "./diaryEntryMealType";

export interface DiaryEntry {
  id: number;
  date: Date;
  mealType: DiaryEntryMealType;
  foodId?: number;
  foodName: string;
  quantityG: number;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}