# lib/api-zod/src/generated/types/food.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriControl - Nutritional tracking app API
 * OpenAPI spec version: 0.1.0
 */

export interface Food {
  id: number;
  name: string;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
  fiberG?: number;
  servingG: number;
}