# lib/api-zod/src/generated/types/hydration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriControl - Nutritional tracking app API
 * OpenAPI spec version: 0.1.0
 */

export interface Hydration {
  date: Date;
  totalMl: number;
  cups: number;
}