# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NutriControl - Nutritional tracking app API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileActivityLevel } from "./profileActivityLevel";
import type { ProfileCaloricGoal } from "./profileCaloricGoal";
import type { ProfileSex } from "./profileSex";

export interface Profile {
  id: number;
  name: string;
  age: number;
  sex: ProfileSex;
  weightKg: number;
  heightCm: number;
  activityLevel: ProfileActivityLevel;
  caloricGoal: ProfileCaloricGoal;
  bmi: number;
  bmr: number;
  tdee: number;
  targetCalories: number;
  targetProteinG: number;
  targetCarbsG: number;
  targetFatG: number;
  targetWaterMl: number;
  intermittentFasting: boolean;
  fastingStartHour?: number;
  fastingEndHour?: number;
}